split structured types integration tests
